refactor: Split player + skill row components from their data
