+++ src/components/SkillBreakdown.tsx
@@ -1,0 +1,43 @@
+import { ComputedPlayerState } from "@/types";
+
+import { SkillRow } from "./SkillRow";
+import { useSkillBreakdown } from "./useSkillBreakdown";
+
+export type SkillBreakdownProps = {
+  player: ComputedPlayerState;
+  color: string;
+};
+
+export const SkillBreakdown = ({ player, color }: SkillBreakdownProps) => {
+  const { skills } = useSkillBreakdown(player);
+
+  return (
+    <tr className="skill-table">
+      <td colSpan={100}>
+        <table className="table w-full">
+          <thead className="header transparent-bg">
+            <tr>
+              <th className="header-name">Skill</th>
+              <th className="header-column text-center">Hits</th>
+              <th className="header-column text-center">Total</th>
+              <th className="header-column text-center">Min</th>
+              <th className="header-column text-center">Max</th>
+              <th className="header-column text-center">Avg</th>
+              <th className="header-column text-center">%</th>
+            </tr>
+          </thead>
+          <tbody className="transparent-bg">
+            {skills.map((skill) => (
+              <SkillRow
+                key={`${skill.childCharacterType}-${skill.groupName}`}
+                characterType={player.characterType}
+                skill={skill}
+                color={color}
+              />
+            ))}
+          </tbody>
+        </table>
+      </td>
+    </tr>
+  );
+};
