+++ src/components/SkillRow.tsx
@@ -1,0 +1,83 @@
+import { CharacterType, ComputedSkillState } from "@/types";
+import { useSkillRow } from "./useSkillRow";
+
+export type SkillRowProps = {
+  characterType: CharacterType;
+  skill: ComputedSkillState;
+  color: string;
+};
+
+export const SkillRow = ({ skill, color }: SkillRowProps) => {
+  const {
+    showFullValues,
+    totalDamage,
+    totalDamageUnit,
+    minDmg,
+    minDmgUnit,
+    maxDmg,
+    maxDmgUnit,
+    rawAverageDmg,
+    averageDmg,
+    averageDmgUnit,
+  } = useSkillRow(skill);
+
+  return (
+    <tr className="skill-row">
+      <td className="text-left row-data">{skill.groupName}</td>
+      <td className="text-center row-data">{skill.hits}</td>
+      <td className="text-center row-data">
+        {showFullValues ? (
+          skill.totalDamage.toLocaleString()
+        ) : (
+          <>
+            {totalDamage}
+            <span className="unit font-sm">{totalDamageUnit}</span>
+          </>
+        )}
+      </td>
+      <td className="text-center row-data">
+        {showFullValues ? (
+          skill.minDamage ? (
+            skill.minDamage.toLocaleString()
+          ) : (
+            ""
+          )
+        ) : (
+          <>
+            {skill.minDamage && minDmg}
+            <span className="unit font-sm">{minDmgUnit}</span>
+          </>
+        )}
+      </td>
+      <td className="text-center row-data">
+        {showFullValues ? (
+          skill.maxDamage ? (
+            skill.maxDamage.toLocaleString()
+          ) : (
+            ""
+          )
+        ) : (
+          <>
+            {skill.maxDamage && maxDmg}
+            <span className="unit font-sm">{maxDmgUnit}</span>
+          </>
+        )}
+      </td>
+      <td className="text-center row-data">
+        {showFullValues ? (
+          rawAverageDmg.toLocaleString()
+        ) : (
+          <>
+            {averageDmg}
+            <span className="unit font-sm">{averageDmgUnit}</span>
+          </>
+        )}
+      </td>
+      <td className="text-center row-data">
+        {skill.percentage.toFixed(0)}
+        <span className="unit font-sm">%</span>
+      </td>
+      <div className="damage-bar" style={{ backgroundColor: color, width: `${skill.percentage}%` }} />
+    </tr>
+  );
+};
